Adding inputset sub-package
inputset provide general interface for input templating. Templates
are supplied as yaml files. Additional templates can be stored at
~/.inputset although those in the source folder will take the
precedence. Note that these templates are only for convenience,
and exactly which is used will  NOT be recorded, and whole provenance
of the exact inputs is always recorded. It is possible to edit
them as one like.