Added function to get parallelisation strategy
This function can be used by the workchains to find the best
parallelisation settings (KPAR, NCORE). For KPAR, the default is
to maximise it within the constraint that the expected wavefunction
size per process cannot exceed a certain limit (default to 1000MB).
For NCORE, the default is to maximise it under the constraint that
it should be less than 12 as well as being factor of the process
counts per node. This might not be ideal for massive parallel machines,
there the "balance" strategy should be used to balance between the
NPAR and NCORE.