Fix a bug for AutoPhononWorkChain
Previously the initial unrelaxed structure was used to ad-hoc phonon
calculation instead of the high quality relaxed structure. This can
result in incorrect results in the computed mesh/band structure.
All results taken before the `0.1.0` version are affected.

However, it is possible to reconstruct the phonon object from the
outputs of the workchain for future analysis or export the calculations
and work on with the command line. It was just the direct outputs of
these workchains are incorrect (thermal properties/band_structure).