Make final singlepoint calculation respect the reuse
This can be important for magnetisation calculations where the
final spin state may not be the same as the initial magmom.
Hence resusing the folder is one of the reliable way of continuting
the calculation.
In the future, one should alled the project magnetisation to be
carried over as an alternative. This also avoiding having too
may intermediate files.