Initial implementation of the hybrid bs wc
A seperate workchain is added to do band structures for hybrid
calculations. This involves splitting the kpoint pathways into
multiple segments and add them to normal SCF calculation but
with zero weights. The workchain then combines the extra kpoints
into a `BandsData` node.

However, this requires a fully configured `KpointData` to be passed
for the SCF calculation with explicity points and weights. This
can be obtained by dryrunning a calculation beforehand, or read
from an existing IBZKPT file/vasprun.xml results.