Added the function for extracting kpoints from finished calcualtions
This allows extracting kpoints used for a finished calculation by
looking into the vasprun.xml file. Compressed files are not supported
for now.

In the future, should write a general method for extracting the
retrieved contant into a temporary directory and file handling,
since disk-based access will be removed in aiida 2.0.