Fix handling of unfinished relaxation
Add the handling of unfinished calculations. This was resulting in
zero return code previously, but now will give an error code in
recent versions of aiida-vasp.

In addtion, add checks for final residual force present in the
SCF calculation. Check should be skipped if the final singlepoint
calculation is performed using the tetrahedral method for BZ integration.